refactor(etfs): replace dynamic ETF sampling with a hardcoded 16-ETF JSON
Drop the /diverse-etfs endpoint, the sample-etfs.ts script, the
etf-verify:sample yarn command, and the most-famous-etfs-by-group.json
seed data. Replace them with a single hardcoded list at
tasks/koala-gains/etf-verification/sample-etfs.json — 2 ETFs per group
in different fund categories, 16 total. Claude just points
trigger-generation and fetch-analysis at that file; there is no
sampling step anymore.

Also moved the SampledEtf type into lib.ts so the two remaining scripts
don't have to cross-import from a deleted file, and updated the
short / long runbooks to reflect the new 4-step flow (set env, trigger,
wait, fetch + review).
